Drop `workflow_dispatch` build method
It causes unnecessary complication: because `python -m build` needs `git tag` to have been run already, so we needed to awkwardly both run that ourselves and then feed it into action-gh-release which redoes the tagging.

Instead, make only one way to do it: click to the Releases page and make a release from there, which guarantees the tag to be pre-created.