Make sure to log to syslog.
This isn't exactly on topic since this detail *should* be taken care
of by a packaged version you can pkg_add, but until then this is
important info.

Connecting stderr to syslog was only added to OpenBSD recently, so mea
culpa. At the moment this just replicates the way rc.subr(8) does it,
but with a `&` added to trigger daemonizing.

See https://github.com/Kozea/Radicale/pull/1194#issuecomment-1104527318